feat: geosort and geofilter return all types
