+++ backend/src/routes/geolocation(raw).js
@@ -4,7 +4,7 @@ const geolocationController = require("../controllers/geolocation.js");
 const validateToken = require("../middlewares/validateToken.js");
 const getIpLocation = require("../middlewares/ipLocation.js");
 
-router.get("/sort", getIpLocation, geolocationController.sortDistanceAll);
-router.get("/filter", getIpLocation, geolocationController.filterDistanceAll);
+router.get("/sort", getIpLocation, geolocationController.sortDistance);
+router.get("/filter", getIpLocation, geolocationController.filterDistance);
 
 module.exports = router;
